deployments/platformdeployment/Docker/bin/requirements.txt
boto3[crt]
pyyaml
pydig
